api: create a v2 endpoint for creating backups
