+++ api/views/mixins.py
@@ -1,6 +1,9 @@
+from inspect import cleandoc
 from drf_spectacular.utils import extend_schema
 from rest_framework import status
 from rest_framework.response import Response
+from rest_framework.views import APIView
+from lib import backup_helper
 from lib.auth import authorized
 
 from anthias_app.models import Asset
@@ -23,3 +26,28 @@ class DeleteAssetViewMixin:
         asset.delete()
 
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+class BackupViewMixin(APIView):
+    @extend_schema(
+        summary='Create backup',
+        description=cleandoc("""
+        Create a backup of the current Anthias instance, which
+        includes the following:
+        * current settings
+        * image and video assets
+        * asset metadata (e.g. name, duration, play order, status),
+          which is stored in a SQLite database
+        """),
+        responses={
+            201: {
+                'type': 'string',
+                'example': 'anthias-backup-2021-09-16T15-00-00.tar.gz',
+                'description': 'Backup file name'
+            }
+        }
+    )
+    @authorized
+    def post(self, request):
+        filename = backup_helper.create_backup(name=settings['player_name'])
+        return Response(filename, status=status.HTTP_201_CREATED)
